+++ src/main/java/ua/hudyma/domain/spells/enums/WaterSpellSchool.java
@@ -17,7 +17,7 @@ public enum WaterSpellSchool implements AbstractSpellSchool {
     private final Integer manaCost;
     private final PrimarySkill spellPrimarySkill; // hero's primary skill accounted in spell damage calculation (mostly WISDOM)
     private final Integer modifierCoefficient;
-    private final List<Integer> modifiedValuesList;
+    private final List<Float> modifiedValuesList;
 
     @Override
     public int getSpellLevel() {
@@ -50,7 +50,7 @@ public enum WaterSpellSchool implements AbstractSpellSchool {
     }
 
     @Override
-    public List<Integer> getModifiedValuesList() {
+    public List<Float> getModifiedValuesList() {
         return modifiedValuesList;
     }
 }
